+++ server.py
@@ -7,7 +7,10 @@ import json
 from multiprocessing.connection import wait
 from socket import *
 import csv
+from xml.etree.ElementPath import find
 import numpy as np
+import RFW_pb2
+import RFD_pb2
 
 def find_avg(samples):
     length=len(samples)
@@ -31,7 +34,13 @@ def max_min(samples,options):
 
 def find_percentile(samples,DA):
     samples.sort()
-    return np.percentile(np.array(np.double(samples)),int(DA))
+    temp=list(DA)
+    temp2=''
+    p = temp.index("p")  # find position of the letter "a"
+    del(temp[p])
+    for element in temp: 
+        temp2 += str(element)
+    return np.percentile(np.array(np.double(samples)),int(temp2))
 
 def find_std(samples):#Finding as if samples were the population
     samples.sort()
@@ -48,31 +57,45 @@ space=" "
 
 print('Server ready to receive on port:'+str(serverPort))
 while True:
+    f,clientAddress=serverSocket.recvfrom(2048)
+    format=f.decode()
+
     message,clientAddress=serverSocket.recvfrom(2048)
     mM=message.decode()
     
     #
     if mM!="":
-
-        filename=mM
-        with open("Server/"+filename, 'w') as f:
-            data = serverSocket.recv(4096)
-            f.write(data.decode())
-            print(f.name+" has been downloaded successfully.")
-        
-        #variable assignment
-        with open('Server/RFW.json', 'r') as file:
-            data=json.load(file)
-            id=data['RFW']["ID"]
-            BenchmarkType=data['RFW']["BenchmarkType"]
-            WorkloadMetric=data['RFW']["WorkloadMetric"]
-            BatchUnit=data['RFW']["BatchUnit"]
-            BatchID=data['RFW']["BatchID"]
-            BatchSize=data['RFW']["BatchSize"]
-            DataType=data['RFW']["DataType"]
-            DataAnalytics=data['RFW']["DataAnalytics"]
+        if(format=="j"):
+            with open("Server/RFW.json", 'w') as f:
+                data = serverSocket.recv(4096)
+                f.write(data.decode())
+                print(f.name+" has been downloaded successfully.")
             
-        file.close()
+            #variable assignment
+            with open('Server/RFW.json', 'r') as file:
+                data=json.load(file)
+                id=data['RFW']["ID"]
+                BenchmarkType=data['RFW']["BenchmarkType"]
+                WorkloadMetric=data['RFW']["WorkloadMetric"]
+                BatchUnit=data['RFW']["BatchUnit"]
+                BatchID=data['RFW']["BatchID"]
+                BatchSize=data['RFW']["BatchSize"]
+                DataType=data['RFW']["DataType"]
+                DataAnalytics=data['RFW']["DataAnalytics"]
+                
+            file.close()
+        else:
+            test = RFW_pb2.Rfw()
+            test.ParseFromString(message)
+            #variable assignment
+            id=test.ID
+            BenchmarkType=test.BenchmarkType
+            WorkloadMetric=test.WorkloadMetric
+            BatchUnit=test.BatchUnit
+            BatchID=test.BatchID
+            BatchSize=test.BatchSize
+            DataType=test.DataType
+            DataAnalytics=test.DataAnalytics
         
         #file processing
         
@@ -101,16 +124,21 @@ while True:
             for row in csv_reader:
                 if rowcount==line_count:
                     if line_count<condition:
-                        print(f'Before: {count}')
                         data_samples[count]=np.double(row[Colnum])
-                        print(data_samples)
                         line_count+=1
                         count+=1
                     else:
                         break
                 rowcount+=1
-            print(f'Lines processed: {line_count}')
-            print(str(data_samples))    
+            print(f'Lines processed: {line_count}') 
+            
+            counter=0
+            d1=[0]*(len(data_samples))
+            for x in data_samples:
+                d1[counter]=data_samples[counter]
+                counter+=1
+            
+
         #Calculate analytics
               
 
@@ -126,31 +154,42 @@ while True:
         else:
             data_analytic=find_percentile(data_samples,DataAnalytics)
         
-        
-        #Write to RFD.json
-        with open('Server/RFD.json', 'r+') as f:
-            data = json.load(f)
-            data['RFD']["ID"]=id
-            data['RFD']["LastBatchID"]=str(last_batch_id)
-            data['RFD']["DataSamples"]=data_samples
-            data['RFD']["DataAnalytic"]=str(data_analytic)
-            print("JSON file updated")
-            f.seek(0)        
-            json.dump(data, f, indent=1)
-            f.truncate()     
+        if(format=="j"):
+            #Write to RFD.json
+            with open('Server/RFD.json', 'r+') as f:
+                data = json.load(f)
+                data['RFD']["ID"]=id
+                data['RFD']["LastBatchID"]=str(last_batch_id)
+                data['RFD']["DataSamples"]=d1
+                data['RFD']["DataAnalytic"]=str(data_analytic)
+                print("JSON file updated")
+                f.seek(0)        
+                json.dump(data, f, indent=1)
+                f.truncate()     
 
 
-        #Send RFD.json to client
-        filename="RFD.json"
-        #serverSocket.sendto(filename.encode(),clientAddress)
-        f = open("Server/"+filename,'rb')
-        l = f.read(2048)
-        while (l):
-            serverSocket.sendto(l,clientAddress)
-            print('Sending...')
+            #Send RFD.json to client
+            filename="RFD.json"
+            #serverSocket.sendto(filename.encode(),clientAddress)
+            f = open("Server/"+filename,'rb')
             l = f.read(2048)
-        f.close()
-        print('Done sending')
+            while (l):
+                serverSocket.sendto(l,clientAddress)
+                print('Sending...')
+                l = f.read(2048)
+            f.close()
+            print('Done sending')
+        else:
+            #Set Values for RFD
+            RFD = RFD_pb2.Rfd()
+            RFD.ID=id
+            RFD.LastBatchID=str(last_batch_id)
+            RFD.DataSamples.extend(d1)
+            RFD.DataAnalytic=str(data_analytic)
+
+            #Send back to client
+            serverSocket.sendto(RFD.SerializeToString(),clientAddress)
+            print('Message Sent!')
 
     #Not a valid command
     else:
